+++ NEWS.md
@@ -31,6 +31,8 @@
 
 - Support `encoding` option in `ReadCharacters()` function family.
 
+- Support `CHARSTATELABELS` in `ReadCharacters()`.
+
 - Support for more formatting quirks in `ReadNotes()`.